重构模块：重命名共享类避免冲突，新增bootstrap启动模块，精简pom依赖，优化前端页面
- 将 GlobalExceptionHandler/MyBatisPlusConfig/TraceCodeMapper 重命名为模块专属名称，规避多模块冲突
- 新增 bootstrap 启动模块，统一管理多模块应用启动
- 精简各模块 pom.xml 依赖项
- 优化前端 Dashboard/ColdChain/LoginRegister/Warnings 页面
- 调整测试配置文件及测试用例
- 添加 Maven Wrapper 和根 pom.xml

+++ enterprise/src/test/java/com/foodtraceability/enterprise/ColdChainServiceSimpleTest.java
@@ -25,12 +25,12 @@ public class ColdChainServiceSimpleTest {
     private ColdChainService coldChainService;
 
     // 用于跨测试方法共享数据
-    private static Integer warehouseId;
+    private static Long warehouseId;
     private static String warehouseUuid;
     private static String warehouseName = "测试冷库A";
-    private static Integer vehicleId;
+    private static Long vehicleId;
     private static String plateNo = "测A·TEST01";
-    private static Integer transportId;
+    private static Long transportId;
     private static String orderNo;
 
     // ==================== t_warehouse ====================
